% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bottom_fishing_areas.R
\docType{data}
\name{Bottom_fishing_areas}
\alias{Bottom_fishing_areas}
\title{Bottom fishing areas}
\format{
A SpatialPolygonsDataFrame
}
\source{
\itemize{
\url{https://marine-conservation.org/what-we-do/program-areas/high-seas/deep-sea-benthic-protections/}
\url{http://www.fao.org/in-action/vulnerable-marine-ecosystems/background/vme-tools/en/#c325329}
}
}
\description{
Bottom fishing areas from FAO
}
\details{
This dataset has compiled 5  separate regional fisheries management organizations (RFMOs).
}
\references{
Gianni, M., Fuller, S.D., Currie, D.E.J., Schleit, K., Goldsworthy, L., Pike, B., Weeber,
B., Owen, S., Friedman, A. Deep Sea Conservation Coalition, July 2016.
}
\keyword{datasets}
