% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change.colors.R
\name{change.colors}
\alias{change.colors}
\title{Color palettes to indicate change}
\usage{
change.colors(x)
}
\arguments{
\item{x}{A Raster* object}
}
\description{
This function return a list of color breaks and colors with white color showing no change
and warm and cold colors for the positive and negative changes respectively.
}
\examples{
# Changes in seafloor temperature at 2041 to 2060
col <- change.colors(subset(cimp5_2041_2060_ch,4))
mplot(subset(cimp5_2041_2060_ch,4), at=col$breaks, col.regions=col$col, main= expression(Temperature~(degree~C)))
}
\author{
Chih-Lin Wei <chihlinwei@gmail.com>
}
