% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cimp5_sdtoe_constant.R
\docType{data}
\name{cimp5_sdtoe_constant}
\alias{cimp5_sdtoe_constant}
\title{Model-average Constant Time of Emergence}
\format{A RasterBrick object of 4 raster layers:
\describe{
  \item{epc_sdtoe_constant}{Constant time of emergence in export POC flux to seafloor}
  \item{o2_sdtoe_constant}{Constant time of emergence in dissolved oxygen concentration at seafloor}
  \item{ph_sdtoe_constant}{Constant time of emergence in pH at seafloor}
  \item{thetao_sdtoe_constant}{Constant time of emergence in potential temperature at seafllor}
  \item{arag_sdtoe_constant}{Constant time of emergence in aragonite Concentration}
  \item{calc_sdtoe_constant}{Constant time of emergence in calcite Concentration}
  \item{co3_sdtoe_constant}{Constant time of emergence in mole Concentration of Carbonate expressed as Carbon in Sea Water}
  \item{co3satarag_sdtoe_constant}{Constant time of emergence in carbonate ion concentration for seawater in equilibrium with pure aragonite}
  \item{co3satcalc_sdtoe_constant}{Constant time of emergence in carbonate ion concentration for seawater in equilibrium with pure calcite}
  \item{aragsat_sdtoe_constant}{Earliest time of emergence in aragonite Saturation State}
  \item{calcsat_sdtoe_constant}{Earliest time of emergence in calcite Saturation State}
}}
\source{
\url{https://esgf-node.llnl.gov/search/esgf-llnl/}
}
\description{
Constant time of emergence was estimated as the earliest year in a latest consecutive period
when climate change variability exceeds historical climate variability.
}
\details{
Constant time of emergence was calculated as the earliest in a latest consecutive period
when the yearly cumulative future standard deviation exceed the historical yearly standard deviation.
Gap years (ratio < 1) within the consecutive period are allowed as long as the proportion of gap years
is less than 10% of the total years within the period.
}
\keyword{datasets}
