% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask2500.R
\docType{data}
\name{mask2500}
\alias{mask2500}
\title{ETOPO1 Global Relief Model masked by 200 to 2500 m depth}
\format{A RasterLayer object of water depth}
\source{
\url{https://www.ngdc.noaa.gov/mgg/global/global.html}
}
\description{
A RasterBrick of the Global Relief Model. A multi-layer raster object compiled from the Cell/pixel-registered ETOPO1 bedrock global relief model.
The grid resolution is reduced from 1 minutes to 0.5 degree by averaging.
}
\examples{
mplot(mask2500, col.regions=terrain.colors,
      cut=100,
      main=expression(Depth~(m)))
}
\keyword{datasets}
