% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplot.R
\name{mplot}
\alias{mplot}
\title{A simple mapping function}
\usage{
mplot(x, loc = NULL, fill = "transparent", col = "gray30", ...)
}
\arguments{
\item{x}{A Raster* object}

\item{loc}{A SpatialPointDataFrame object to add to the map}

\item{...}{Any argument that can be passed to \code{\link{spplot}} and \code{\link{levelplot}}}
}
\description{
This function is a wrapper of  \code{\link{spplot}}
}
\examples{
data("woa13ts")
mplot(woa13ts, col.regions=jet.colors2, cut=100)
}
\author{
Chih-Lin Wei <chihlinwei@gmail.com>
}
