% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scale_colorPalette.R
\name{scale_colorPalette}
\alias{scale_colorPalette}
\title{Re-scale Color Palettes Function}
\usage{
scale_colorPalette(col_fun = jet.colors, scale = "2nd")
}
\arguments{
\item{col_fun}{A Color Palettes function}

\item{scale}{A character string of "2nd", "4th" and "log" to rescale the color palettes indicated by col_fun}
}
\value{
returns a function that takes an integer argument (the required number of colors) and returns a character vector of colors (see \code{\link{rgb}}) interpolating the given sequence (similar to \code{\link{heat.colors}} or \code{\link{terrain.colors}}.
}
\description{
This function is a wrapper of color interpolation  \code{\link{colorRampPalette}},
which takes an color palettes function
and re-scale palettes in square root, 4th root and log scales.
}
\examples{
log.colors <- scale_colorPalette(heat.colors, "log")
}
\author{
Chih-Lin Wei <chihlinwei@gmail.com>
}
