% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seamount.R
\docType{data}
\name{seamount}
\alias{seamount}
\title{Global distribution of large seamounts}
\format{
A SpatialPointsDataFrame
}
\source{
\url{https://doi.org/10.1111/j.1365-246X.2011.05076.x}
}
\description{
Global distribution of large seamounts from Kim and Wessel (2011)
}
\details{
This dataset has compiled 24643  separate large seamounts in the world ocean.
}
\examples{
# VME-associated seamounts
# Seamount submmits between 200 and 2500 m water depth
library(plotKML)
seamount@data$Summit <- seamount@data$Height+seamount@data$Depth
plotKML(subset(seamount, Summit>=-2500 & Summit<=-200)[, "Summit"],
        folder.name="seamount", colour_scale = jet.colors(100), raster_name="seamount.png")
}
\references{
Kim, S.-S., Wessel, P., 2011. New global seamount census from altimetry-derived gravity data.
Geophysical Journal International 186, 615–631.
}
\keyword{datasets}
